+++ src/provider/index.ts
@@ -5,6 +5,7 @@ export { default as BingProvider } from "provider/BingProvider";
 export * from "provider/BingProvider";
 export { default as ChainProvider } from "provider/ChainProvider";
 export * from "provider/ChainProvider";
+export * from "provider/geoplugin";
 export * from "provider/googlemaps";
 export * from "provider/mapbox";
 export { default as MapquestProvider } from "provider/MapquestProvider";