+++ src/query/GeocodeQuery.ts
@@ -2,7 +2,8 @@ import { DEFAULT_RESULT_LIMIT } from "provider";
 import { Bounds } from "index";
 
 export interface GeocodeQueryObject {
-  readonly text: string;
+  readonly text?: string;
+  readonly ip?: string;
   readonly south?: number | string;
   readonly west?: number | string;
   readonly north?: number | string;
@@ -12,7 +13,9 @@ export interface GeocodeQueryObject {
 }
 
 export default class GeocodeQuery {
-  private readonly text: string;
+  private readonly text?: string;
+
+  private readonly ip?: string;
 
   private readonly south?: number | string;
 
@@ -28,6 +31,7 @@ export default class GeocodeQuery {
 
   protected constructor({
     text,
+    ip,
     south,
     west,
     north,
@@ -36,6 +40,10 @@ export default class GeocodeQuery {
     limit = DEFAULT_RESULT_LIMIT,
   }: GeocodeQueryObject) {
     this.text = text;
+    this.ip = ip;
+    if (!text && !ip) {
+      throw new Error('Either "text" or "ip" parameter is required.');
+    }
     this.south = south;
     this.west = west;
     this.north = north;
@@ -51,6 +59,7 @@ export default class GeocodeQuery {
   public toObject(): GeocodeQueryObject {
     return {
       text: this.text,
+      ip: this.ip,
       south: this.south,
       west: this.west,
       north: this.north,
@@ -64,6 +73,13 @@ export default class GeocodeQuery {
     return (<typeof GeocodeQuery>this.constructor).create({
       ...this.toObject(),
       text,
+    });
+  }
+
+  public withIp(ip: string): GeocodeQuery {
+    return (<typeof GeocodeQuery>this.constructor).create({
+      ...this.toObject(),
+      ip,
     });
   }
 
@@ -96,8 +112,12 @@ export default class GeocodeQuery {
     });
   }
 
-  public getText(): string {
+  public getText(): undefined | string {
     return this.text;
+  }
+
+  public getIp(): undefined | string {
+    return this.ip;
   }
 
   public getBounds(): undefined | Bounds {