+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0
 
 * Add special chain provider
+* Add GeoPlugin provider
 * Add `errorCallback` parameter to `geocode` and `geodecode`
 * Add `body` parameter to `executeRequest`
 * [MapQuest] Fix geocode request (bad encoded query)